+++ Sources/BraintreeShopperInsights/V2/BTCustomerSessionRequest.swift
@@ -13,7 +13,7 @@ public struct BTCustomerSessionRequest {
     let payPalAppInstalled: Bool?
     let venmoAppInstalled: Bool?
     let purchaseUnits: [BTPurchaseUnit]?
-    let payPalCampaigns: [BTPayPalCampaign]
+    let campaigns: [BTPayPalCampaign]
     
     /// Creates a BTCustomerSessionRequest
     /// - Parameters:
@@ -22,7 +22,7 @@ public struct BTCustomerSessionRequest {
     ///   - payPalAppInstalled: Optional: Checks whether the PayPal app is installed on the device.
     ///   - venmoAppInstalled: Optional: Checks whether the Venmo app is installed on the device.
     ///   - purchaseUnits: Optional: The list of purchase units containing the amount and currency code.
-    ///   - payPalCampaigns: Optional: PayPal campaigns displayed to the customer in the shopping journey.
+    ///   - campaigns: Optional: PayPal campaigns displayed to the customer in the shopping journey.
     /// - Warning: This feature is in beta. It's public API may change or be removed in future releases.
     public init(
         hashedEmail: String? = nil,
@@ -30,13 +30,13 @@ public struct BTCustomerSessionRequest {
         payPalAppInstalled: Bool? = nil,
         venmoAppInstalled: Bool? = nil,
         purchaseUnits: [BTPurchaseUnit]? = nil,
-        payPalCampaigns: [BTPayPalCampaign] = []
+        campaigns: [BTPayPalCampaign] = []
     ) {
         self.hashedEmail = hashedEmail
         self.hashedPhoneNumber = hashedPhoneNumber
         self.payPalAppInstalled = payPalAppInstalled
         self.venmoAppInstalled = venmoAppInstalled
         self.purchaseUnits = purchaseUnits
-        self.payPalCampaigns = payPalCampaigns
+        self.campaigns = campaigns
     }
 }